Move expansion types out of pkg/common
