+++ cmd/expandybird/expander/expander_test.go
@@ -25,14 +25,15 @@ import (
 
 	"github.com/kubernetes/helm/pkg/chart"
 	"github.com/kubernetes/helm/pkg/common"
+	"github.com/kubernetes/helm/pkg/expansion"
 )
 
 var expanderName = "../../../expansion/expansion.py"
 
 type testCase struct {
 	Description      string
-	Request          *common.ExpansionRequest
-	ExpectedResponse *common.ExpansionResponse
+	Request          *expansion.ExpansionRequest
+	ExpectedResponse *expansion.ExpansionResponse
 	ExpectedError    string
 }
 
@@ -47,8 +48,8 @@ func funcName() string {
 	return runtime.FuncForPC(pc).Name()
 }
 
-func testExpansion(t *testing.T, req *common.ExpansionRequest,
-	expResponse *common.ExpansionResponse, expError string) {
+func testExpansion(t *testing.T, req *expansion.ExpansionRequest,
+	expResponse *expansion.ExpansionResponse, expError string) {
 	backend := NewExpander(expanderName)
 	response, err := backend.ExpandChart(req)
 	if err != nil {
@@ -81,7 +82,7 @@ var jinjaExpander = &chart.Expander{
 func TestEmptyJinja(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -99,7 +100,7 @@ func TestEmptyJinja(t *testing.T) {
 				},
 			},
 		},
-		&common.ExpansionResponse{
+		&expansion.ExpansionResponse{
 			Resources: []interface{}{},
 		},
 		"", // Error
@@ -109,7 +110,7 @@ func TestEmptyJinja(t *testing.T) {
 func TestEmptyPython(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -130,7 +131,7 @@ func TestEmptyPython(t *testing.T) {
 				},
 			},
 		},
-		&common.ExpansionResponse{
+		&expansion.ExpansionResponse{
 			Resources: []interface{}{},
 		},
 		"", // Error
@@ -140,7 +141,7 @@ func TestEmptyPython(t *testing.T) {
 func TestSimpleJinja(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -162,7 +163,7 @@ func TestSimpleJinja(t *testing.T) {
 				},
 			},
 		},
-		&common.ExpansionResponse{
+		&expansion.ExpansionResponse{
 			Resources: []interface{}{
 				map[string]interface{}{
 					"name": "foo",
@@ -177,7 +178,7 @@ func TestSimpleJinja(t *testing.T) {
 func TestSimplePython(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -201,7 +202,7 @@ func TestSimplePython(t *testing.T) {
 				},
 			},
 		},
-		&common.ExpansionResponse{
+		&expansion.ExpansionResponse{
 			Resources: []interface{}{
 				map[string]interface{}{
 					"name": "foo",
@@ -216,7 +217,7 @@ func TestSimplePython(t *testing.T) {
 func TestPropertiesJinja(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -244,7 +245,7 @@ func TestPropertiesJinja(t *testing.T) {
 				},
 			},
 		},
-		&common.ExpansionResponse{
+		&expansion.ExpansionResponse{
 			Resources: []interface{}{
 				map[string]interface{}{
 					"name": "foo",
@@ -262,7 +263,7 @@ func TestPropertiesJinja(t *testing.T) {
 func TestPropertiesPython(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -292,7 +293,7 @@ func TestPropertiesPython(t *testing.T) {
 				},
 			},
 		},
-		&common.ExpansionResponse{
+		&expansion.ExpansionResponse{
 			Resources: []interface{}{
 				map[string]interface{}{
 					"name": "foo",
@@ -310,7 +311,7 @@ func TestPropertiesPython(t *testing.T) {
 func TestMultiFileJinja(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -336,7 +337,7 @@ func TestMultiFileJinja(t *testing.T) {
 				},
 			},
 		},
-		&common.ExpansionResponse{
+		&expansion.ExpansionResponse{
 			Resources: []interface{}{
 				map[string]interface{}{
 					"name": "foo",
@@ -368,7 +369,7 @@ var schemaContent = content([]string{
 func TestSchema(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -401,7 +402,7 @@ func TestSchema(t *testing.T) {
 				},
 			},
 		},
-		&common.ExpansionResponse{
+		&expansion.ExpansionResponse{
 			Resources: []interface{}{
 				map[string]interface{}{
 					"name": "foo",
@@ -419,7 +420,7 @@ func TestSchema(t *testing.T) {
 func TestSchemaFail(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -460,7 +461,7 @@ func TestSchemaFail(t *testing.T) {
 func TestMultiFileJinjaMissing(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -486,7 +487,7 @@ func TestMultiFileJinjaMissing(t *testing.T) {
 func TestMultiFilePython(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -531,7 +532,7 @@ func TestMultiFilePython(t *testing.T) {
 				},
 			},
 		},
-		&common.ExpansionResponse{
+		&expansion.ExpansionResponse{
 			Resources: []interface{}{
 				map[string]interface{}{
 					"name": "foo",
@@ -546,7 +547,7 @@ func TestMultiFilePython(t *testing.T) {
 func TestMultiFilePythonMissing(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -574,7 +575,7 @@ func TestMultiFilePythonMissing(t *testing.T) {
 func TestWrongChartName(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -600,7 +601,7 @@ func TestWrongChartName(t *testing.T) {
 func TestEntrypointNotFound(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -621,7 +622,7 @@ func TestEntrypointNotFound(t *testing.T) {
 func TestMalformedResource(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -650,7 +651,7 @@ func TestMalformedResource(t *testing.T) {
 func TestResourceNoName(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
@@ -679,7 +680,7 @@ func TestResourceNoName(t *testing.T) {
 func TestResourceNoType(t *testing.T) {
 	testExpansion(
 		t,
-		&common.ExpansionRequest{
+		&expansion.ExpansionRequest{
 			ChartInvocation: &common.Resource{
 				Name: "test_invocation",
 				Type: funcName(),
